Fixes for the Pull Request
Silhouetter added, now removed from BaseUnitEntity

although I add to BaseUnitEntity 2 actions to call out:
OnSpriteSet<Sprite> - Called when a new sprite is set to the Unit and sends the new sprite (to be copied by the Silhouette). We don't want BaseUnitEntity to reference the Silouetter at all - so the BaseUnitEntity calls an event which Silouetter can sub to.

OnSpriteFlipX<bool> - Called when the Unit "flips their X" (turning left or right) sending true if flipX is true.